fix(merge-back): move dispose-polyfill to lib/private where perf.ts expects it
