+++ packages/aws-cdk-lib/core/lib/private/dispose-polyfill.ts
