+++ packages/aws-cdk-lib/core/test/fs/fs-fingerprint.test.ts
@@ -3,7 +3,7 @@ import * as os from 'os';
 import * as path from 'path';
 import { FileSystem, IgnoreMode, SymlinkFollowMode } from '../../lib/fs';
 import { contentFingerprint } from '../../lib/fs/fingerprint';
-import '../dispose-polyfill';
+import '../../lib/private/dispose-polyfill';
 
 describe('fs fingerprint', () => {
   describe('files', () => {
